fix data type

+++ week_4_analytics_engineering/taxi_rides_ny/models/staging/stg_green_tripdata.sql
@@ -31,7 +31,7 @@ select
     cast(mta_tax as numeric) as mta_tax,
     cast(tip_amount as numeric) as tip_amount,
     cast(tolls_amount as numeric) as tolls_amount,
-    cast(ehail_fee as integer) as ehail_fee,
+    cast(ehail_fee as numeric) as ehail_fee,
     cast(improvement_surcharge as numeric) as improvement_surcharge,
     cast(total_amount as numeric) as total_amount,
     cast(payment_type as integer) as payment_type,
